extensions for NS_ENUM types in SPM context (*)
Since NS_ENUM types are not being translated to Swift enums while in the SPM context, extensions were added to provide equivalent translations. Hopefully, there's a build setting that can be tweaked so we can get rid of these.

This fixes the non-SPM build and switches back to the less verbose dot syntax when referring to these enums.