Fixes compilation in an app target (*)
There appears to be a bug in Xcode regarding the SWIFT_PACKAGE macro. In some cases, the macro is not defined, leading to compilation failures. This defines a `FLEXLAYOUT_SWIFT_PACKAGE` setting that is used in place of `SWIFT_PACKAGE`. While this doesn't solve the problem, it does allow one to add `FLEXLAYOUT_SWIFT_PACKAGE=1` to their app target's `GCC_PREPROCESSOR_DEFINITIONS` build setting to workaround the problem.